requirements.txt
aiofiles==0.7.0
aiosqlite==0.16.1
asgiref==3.4.0
bcrypt==3.2.0
cffi==1.14.5
click==8.0.1
fastapi==0.65.2
h11==0.12.0
iso8601==0.1.14
Jinja2==3.0.1
MarkupSafe==2.0.1
passlib==1.7.4
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pypika-tortoise==0.1.1
python-decouple==3.4
python-multipart==0.0.5
pytz==2021.1
six==1.16.0
starlette==0.14.2
tortoise-orm==0.17.4
typing-extensions==3.10.0.0
uvicorn==0.14.0
